ppc64_cpu: Limit number of CPUs for frequency calculation
The patch limits the number of threads used for CPU frequency calculation
to the minimum of either the actual number of CPUs in the system or
CPU_SETSIZE(currently 1024). In other words, if the number of CPUs in the
system is less than CPU_SETSIZE, a thread is created for each of the CPUs,
otherwise only CPU_SETSIZE threads are created. This is being done because
sched_setaffinity() will only work on CPU_SETSIZE CPUs, anything over the
limit is not supported.

Signed-off-by: Juliet Kim <julietk@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>