lparstat: correct calculation of physc to use tbr
Currently phsyc is calculated using the sampled PURR values from lparcfg and
a derived timebase that lparstat itself calculates. However, both the Timer
Facilities section in Book III of the ISA and PAPR outline that this resource
calcuation is done by computing the ratio of the PURR delta and the Timebase
Register delta for the sampled period.

I've sent a patch upstream to the kernel list to also export the Timebase
Register (tbr) in /proc/ppc64/lparcfg [1]. Use this value for calculating
physc value when running on a kernel that exposes tbr.

[1] https://patchwork.ozlabs.org/patch/1009906/

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>