powerpc-utils: Suppress errors reading kernel files
A problem was observed with the evaluation of the '${devspec} by
the script 'ofpathname'.  The problem occurred during installation
to an nvme device as follows:

02:29:41,669 INF program: Running... ofpathname /dev/nvme0n1p1
02:29:41,922 INF program: /sbin/ofpathname: line 684: warning: command substitution: ignored null byte in input

Detailed examination showed that instances of the device-tree
property "/namespace/name" within one or more of the PCI devices
contained null bytes which were echoed along with the value e.g.

[root device-tree]# cat /proc/device-tree/pci@800000020000029/pci1014,637@0/namespace/name | hexdump -C
00000000  6e 61 6d 65 73 70 61 63  65 00                    |namespace.|
0000000a
[root device-tree]#

This patch provides improves handling of errors within that script,
and by some of the follow-on uses of the tool by the 'bootlist' script.

* Deal with error messages returned sometimes by evaluation
  of ${devspec} in 'ofpathname'.
* Strip embedded nulls from values provided by kernel files.
* Redirect more error messages to /dev/null

Signed-off-by: Michael Bringmann <mwb@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>