ofpathname: Add support for Mac-compatible OF pathnames
On Macintosh systems, OpenFirmware path follow a slightly
different syntax as compared to other OF systems. In particular,
the disk ID is not preceeded by the string "disk" and - on
SATA systems - the path does not contain the BUS ID as
"scsi@ID" but rather the plug ID in the form "@ID".

Signed-off-by: John Paul Adrian Glaubitz <glaubitz@physik.fu-berlin.de>
Reviewed-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>