lparstat: Add support for cpu-hotplug
For optimizing the read of sysfs file, the file descriptors are
assigned at the beginning of the run but they do not change dynamically
in event of a cpu hotplug, resulting under-estimated load average in
case of new cpus onlined or error in case of cpu removed. This can be
solved by checking for the online cpus every interval and re-assign
the file descriptors again with new set of cpu on hotplug event.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>