define _GNU_SOURCE project-wide
This project uses several nonstandard GNU-specific extensions which
are exposed only when _GNU_SOURCE is defined.  Examples include
CPU_SET(3) and related APIs by necessity and asprintf(3) for
convenience.

Currently _GNU_SOURCE is used in a piecemeal fashion, meaning you have
to add it to the source file when you introduce a use of asprintf, for
example. Also, note that it is used incorrectly in a couple of places:
it must appear before any headers are included for it to work reliably.

There's not much justification for the code duplication and lack of
uniformity here. Put '-D _GNU_SOURCE' in AM_CPPFLAGS for a more
consistent experience across the project's C code.

This will cause all C code to be built with _GNU_SOURCE.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>