serv_config: ensure nul-termination of parameter strings
GCC 9.3.1 warns:

In file included from /usr/include/string.h:495,
                 from src/serv_config.c:50:
In function ‘strncpy’,
    inlined from ‘update_value’ at src/serv_config.c:861:4:
/usr/include/bits/string_fortified.h:106:10: warning: ‘__builtin_strncpy’ specified bound 4998 equals destination size [-Wstringop-truncation]

This means that a value string with length 4998 or greater not
including a terminating nul byte would result in this copy leaving the
destination buffer unterminated, which violates the specification for
ibm,set-system-parameter.

In this case we need only reduce the bound by 1.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>