add option --disable-werror to configure
Commit a6929691fd25 ("treat compiler warnings as errors") imposed
-Werror unconditionally in Makefile.am, in order to prevent the
addition of new warnings to the build. While it's important to keep
the project reasonably warning-clean, it's also necessary to provide
downstream packagers and users in a variety of environments more
flexibility in this area.

Keep -Werror by default but make it configurable. The user may specify
--disable-werror at configure time to prevent compiler warnings from
halting the build. This is a reasonably common way to deal with
this (examples include glibc, gdb, json-c).

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>