drmgr: don't treat pmig as unsupported when -m is omitted
If drmgr is invoked for the pmig connector type but the -m option is
not specified get_command() treats the connector type as unsupported.

(1) root @ linux-td1r: /root/powerpc-utils
Dynamic reconfiguration is not supported for connector type "8" on this system
At least one action must be specified

Add a case block to catch DRC_TYPE_MIGRATION connector type.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>