drmgr: query in-kernel dlpar support by DRC connector type
Support to initiate kernel facilitated dlpar for CPU and MEM was
introduced into drmgr back in 2016 with commit bbc5fc0 however, this
proved problematic for CPU as most distros didn't have in kernel CPU
dlpar support yet. As a result CPU support was disabled with commit
a97c566.

The main issue is that drmgr blindly tests for the existence of
/sys/kernel/dlpar to identify in kernel dlpar support. This test doesn't
reveal if support goes beyond MEM and includes CPU. This further
complicates adding new connector types such as PHB, PCI, and SLOT down
the road, and backwards compatability of newer drmgr on older kernels.

The /sys/kernel/dlpar attribute was extended by the following kernel
commit torvalds/linux@673bc43 that introduced a comma separated list of
device support.

Update drmgr utility to use this mechanism to determine which
device/connector/drc_types are supported by in kernel dlpar.

1.) For the case where /sys/kernel/dlpar does not exist use legacy dlpar
facilities

2.) For the case where /sys/kernel/dlpar exists but does not support
reading we assume only memory support for in kernel and use legacy for
all other types.

3.) For the case where /sys/kernel/dlpar exists and is readable we parse
the returned list for supported devices and fall back to legacy
facilities for anything not in the list.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>