ppc64_cpu: Remove smt_snooze_delay
The kernel has not supported setting the snooze delay through this API
since at least 2014 (v3.14). The sysfs file is present for
compatibility, but it has no effect.

The lack of this sysfs file means we must use a different method for
ppc64_cpu to detect the smt state of the box. It now uses the
threads_per_cpu count from the device tree.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>