man: add manpage for the drmgr utility
Historically we have discouraged users from invoking drmgr from the commandline
outside of PCI hotplug, and as such provided no documentation. However, for the
sake of those brave powerusers willing to risk putting their system in a state
inconsistent from that of theassociated managment interface lets add a manpage
with the basic documentation of the various usage options for each of the
dynamic reconfiguration connector types.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>