Allow hcnqrydev return success when sr-iov VFs are not active
When the sr-iov VF nextwork interfact in the bond is not up, that means
it is failover to the backup virtual network device already. We should
allow hcnqrydev return success in this case, in other words, means it is
safe to dlpar remove sr-iov VFs.

Also add more system and network status info in show_hcnstatus() and move it to
the end of main function.

Signed-off-by: Mingming Cao <mmc@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>