ofpathname: Remove the dependency on bc
The only use of bc is to convert between decimal and hexadecimal
integers, but this functionality is supported by POSIX shell
arithmetic expansion and a POSIX-compatible printf program.  Since
this script uses #!/bin/bash, the printf utility is a bash builtin,
so it won't even add a dependency on an external printf program
(which is provided by the required coreutils package anyway).

Signed-off-by: David Michael <fedora.dm0@gmail.com>