ofpathname: Move definition of SYS_PATH from l2of_vs() to l2of_scsi()
SYS_PATH, which is required to calculate the plug ID of a SCSI/SATA
host in l2of_scsi(), is actually never set in this function but in
l2of_vd() where it is not used at all. Thus, move the definition of
SYS_PATH from l2of_vd() to l2of_scsi() to fix the calculation of the
plug ID in l2of_scsi().

Fixes: 3fb2c44e22 ("ofpathname: Add support for the plug ID of a SCSI/SATA host")
Signed-off-by: John Paul Adrian Glaubitz <glaubitz@physik.fu-berlin.de>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>