Avoid using ifcfg file in checking bonding interface exist status
When configure migratable sr_iov into hybrid network, it checks if
there is existing HNV using the presense of ifcfg file location. This
is not prefered as the location can be different on distros.
This patch fix this with consistantly using NetworkManager nmcli as the
rest of code.

Also this patch fix the remove hybrid network issue with xargs.

Signed-off-by: Mingming Cao <mmc@linux.vnet.ibm.com>