Fix ofpathname race with udev rename
When hcncfgdrc is called to configure an sr-iov or a backend vdice, drmgr
has already returned from DR add. However sometimes the two commands happens too
fast that the OS has not complete its work to get the device ready.
We had to put some wait time to wait for sysfs is ready. However
there is still small race window between udev rename and ofpathname lookup
which leads to ofpathname grabbed a staled devname
while udev is in the middle of rename device to something else.
This cause hcncfgdrc failed to enslave the sr-iov or backend vdevice silently.
And later causing Live Parition Migration failed due to missing backend slave
for primary sr-iov device

First we need to catch the failure of enslave vdevice at the time of
calling do_config_vdevice to enslave the device, second, in the case of failure
waiting for udev settle to complete udev events, then try ofpathename and enslave
again.

Signed-off-by: Mingming Cao <mmc@linux.vnet.ibm.com>