hcnmgr: Fix ofpathname race with udev rename
When hcncfgdrc is called to configure an sr-iov or a backend vdevice, drmgr
has already returned from DR add. However sometimes the two commands happens too
fast that the OS has not completed its work to get the device ready. There is
still a small race window between udev rename and ofpathname lookup. This leads
to ofpathname grabbing a stale devname as a result of udev renaming a device to
something else. This causes hcncfgdrc failure to enslave the sr-iov or backend
vdevice silently, and later causes Live Parition Migration failure due to
missing backend slave for primary sr-iov device

Fix by catching the failure of enslave vdevice at the time of calling
do_config_vdevice to enslave the device, and in the case of failure wait for
udev settle to complete udev events before repeating enslave attempt.

Signed-off-by: Mingming Cao <mmc@linux.vnet.ibm.com>
[tyreld: reworded change log]
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>