ofpathname: Add missing substring extraction of devpart in l2of_vd()
l2of_vd() contains the necessary bits to append the partition number
to the resulting OFPATH, but it does not actually extract the partition
number from the logical device path in the first place. This adds the
missing substring extraction of the partition number from the logical
device path so that the partition number is appended to OFPATH when
the logical device is a partition.

Signed-off-by: John Paul Adrian Glaubitz <glaubitz@physik.fu-berlin.de>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>