Introducing lparnumascore
This command is computing the LPAR NUMA score for CPU and memory.

After a LPM, resources like CPU and memory could be reassigned to a
different NUMA node. But the Linux kernel is not able to replicate these
changes and since the kernel commit bb7c3d36e3b1 ("powerpc/pseries: remove
prrn special case from DT update path"), PRRN event are only updating the
device tree, but not the resource's assignement the kernel is managing.

This new command detects CPU and memory resources the kernel keep binded to
the NUMA node topology of the departure node. It computes a ratio, and
print it to the standard output.

The ratio is between 0 (worst) and 100 (best).

For instance when half of the CPUs have been reassigned to a new NUMA node,
the score would be 50. If all the CPUs are on the right node, the score
would be 100.

The command can compute the CPU and memory score or only one if specified
through the -c option.

Example of output:
$ ./lparnumascore -h
Usage: lparnumascore [-d detail_level] [-c {mem | cpu}]

$ ./lparnumascore
CPU score: 80
MEM score: 45

$ ./lparnumascore -c mem
MEM score: 45

$ ./lparnumascore -c cpu
CPU score: 80

The command is not locking /var/lock/dr_config_lock so user should ensure
that there is not ongoing command impacting the computed scores. This
allows the command to be run without special privileges.

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>