bootlist: Fix for multipath devices with > 5 paths
Only five entries can be added to the bootlist. If more than that
are specified, the script flags this as an error and aborts.
However, if a multipath device with more than 5 paths is
specified, this is also failing. Rather than failing to set
any boot entries in this case, we should simply truncate the
boot list to the first 5 paths, which is what this patch does.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>