lsdevinfo: optimize criteria filtering
lsdevinfo is significantly slower when a criteria filter is set with the
-q flag.

There are two culprits:

1. The criteria string given on the command line is parsed every time
   we call check_criteria().  We're forking two sed(1) processes whenever
   we hit that function.  We hit it a lot.

2. Criteria checking runs in constant time.  We call check_criteria() and
   do the parsing in (1) for every relevant attribute, even if we already
   have a match from a prior check_criteria() invocation.

We can fix issue (1) by parsing the criteria string once at the start
of the script.  I have added a function, parse_criteria(), that parses
the $criteria string and selects an appropriate matching function.

This approach also fixes problem (2), but we first need to check
whether the user's criteria is relevant to the attributes the script
cares about before calling the matching function.  We do this with
criteria_is_relevant().

The speed improvement is nice.  Consider my test machine with around
thirty devices.  On this machine, lsdevinfo without any options runs
in 1.35s, but with the "status=1" criteria it runs in 2.12s:

$ /usr/sbin/lsdevinfo | fgrep -c device:
29
$ command time -p /usr/sbin/lsdevinfo > /dev/null
real 1.35
user 1.42
sys 0.16
$ command time -p /usr/sbin/lsdevinfo -q status=1 > /dev/null
real 2.12
user 2.22
sys 0.30

With this patch, lsdevinfo with the "status=1" criteria now runs in
1.35s:

$ command time -p /usr/sbin/lsdevinfo -q status=1 > /dev/null
real 1.35
user 1.41
sys 0.18

This patch eliminates nearly all of the criteria-checking overhead in
the current code.

Signed-off-by: Scott Cheloha <cheloha@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>