hcnmgr: Remove some dead code
There is VNIC_SPT which enables support for vnic as beckup device which
is set based on value of VERSION variable which is hardcoded in the
script and does not correspond the actual HNV support level on the LPAR.

Remove all this. We do support vnic backend so enable it
unconditionally.

It is the job of HMC to decide if vnic backup device is supported and
can be configured. When it is configured by HMC make use of it.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>
Reviewed-by: Mingming Cao <mmc@linux.vnet.ibm.com>
Tested-by: Mingming Cao <mmc@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>