ci: add GitHub Actions build tests
GitHub's hosted x86 Ubuntu runners have cross toolchains suitable for
building powerpc-utils and its various library dependencies (librtas,
zlib, and libnuma) for its targeted architectures. Add a matrix build
including these (32- and 64-bit big-endian, 64-bit little-endian) powerpc
platforms which perform a basic build and distcheck.

Also perform a native x86 build since it "works": it produces nothing you
would want to run, but the build succeeds. If we are to add tests in the
future, running them as native code on the build host may be the best
option. So it's in our interest to keep the x86 build working.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>