drmgr/pmig: remove deprecated call to refrsrc IBM.ManagementServer
As Nathan reported while reviewing this series:

This RSCT doc indicates that the IBM.ManagementServer resource class has
been superseded:

https://www.ibm.com/docs/en/rsct/3.2?topic=security-management-domain-configuration

Removing that deprecated call in this series because it is conflicting with
a following patch. The system returned value is stored in rc to prevent
compilation error (due to warning if system returned value is not read).

Suggested-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>