errinjct: pass full device reg path to get_config_addr_from_reg()
The following string truncation error is reported by gcc-11 and gcc-12
toolchains:

In file included from /usr/powerpc-linux-gnu/include/string.h:535,
	from src/errinjct/ioa_bus_error.c:34:

In function ‘strncpy’,
	inlined from ‘get_config_addr_from_reg’ at src/errinjct/ioa_bus_error.c:207:2,
	inlined from ‘hunt_loc_code’ at src/errinjct/ioa_bus_error.c:415:9:
Warning: /usr/powerpc-linux-gnu/include/bits/string_fortified.h:95:10: warning: ‘__builtin_strncpy’ output may be truncated copying 3995 bytes from a string of length 3999 [-Wstringop-truncation]

This is the result of the caller defining a buffer of BUFSZ, but the
callee only copying BUFSZ-5 of data into a new string so that there is
room to strcat "/reg" to the resulting pathname. We can save a strncpy
and static buffer allocation by doing the strcat of "/reg" to the
pathname before calling get_conig_addr_from_reg() which in general
appears to be inlined anyways.

Reported-by: John Paul Adrian Glaubitz <glaubitz@physik.fu-berlin.de>
Reviewed-by: John Paul Adrian Glaubitz <glaubitz@physik.fu-berlin.de>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>