drmgr/cpu: add CPU add and remove operation's hooks
Call drmgr's hooks before and after a CPU add or remove operation is done.

At the PRE phase, the DRC_COUNT hook parameter is set to the desired number
of CPU to add or remove. At the POST phase, it is set to the actual number
of CPU added or removed.

Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>