nvram: error out if nvram-size is greater than the default value
If the user specifies an nvram-size larger than the default value
(DEFAULT_NVRAM_SZ = 1024 * 1024), it results in a segmentation fault.

Without the patch:

[root@xxx ~]# nvram --nvram-size 268435456
nvram: WARNING: expected 268435456 bytes, but only read 15360!
Segmentation fault (core dumped)

The above issue is fixed by adding a check on nvram-size and error
out if nvram-size is greater than the default value.

with the patch:

[root@xxx src]# ./nvram --nvram-size 268435456
./nvram: ERROR: nvram size must be less than the default nvram-size is 1048576

Signed-off-by: Likhitha Korrapati <likhitha@linux.ibm.com>
Reviewed-by: Sourabh Jain <sourabhjain@linux.ibm.com>