drmgr/phb: Add multipath partner device support for hotplug add
The PHB node can have "ibm,multipath-partner-drc" property
which means the device can be also configured with multipath
partner path. This property provides the DRC index of the
partner device. So for the hotplug add, both paths should be
added and the following steps will be executed for the PHB add:

- Add the specified PHB device
- Find the partner path DRC index from "ibm,multipath-partner-drc"
  property for the specified device
- If the node has this property and the partner path is
  already added, notify user about the partner device.
- If the node has this property and the partner path can be
  configured, notify user to configure and add the partner
  device.

Signed-off-by: Haren Myneni <haren@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>