drmgr/pci: Add multipath partner device support for hotplug remove
If the PCI device has multipath partner device, its device node
contains "ibm,multipath-partner-drc" property which gives the
DRC index of the partner device. For the hotplug removal, if
the partner path is also configured, both paths must be
removed before instructing the user to remove the device from
identified slot.

So the following steps will be executed for the removal:
- Find the partner path DRC index from "ibm,multipath-partner-drc"
  property
- Remove the primary path
- Remove the partner path if it is configured.
- Notify user to remove PCI card from the specified slot.

Since both paths will be using the same slot, LED indicators and
the slot identification will be done only for the primary device.

Signed-off-by: Haren Myneni <haren@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>