ofpathname: skip devices with no devspec when coorelating FC OF paths
When using ofpathname to find the logical device associated to a Fibre
Channel or Virtual Fibre Channel OF path an ERR_NO_SYSFS_DEVINFO error
can be returned if a device with no devspec attribute is in the list of
devices to cross reference.

If a device with no devspec is found simply skip it an continue the
for-loop device search.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>