lparstat: Use pool_capacity for determining active cpus in a pool
The function get_active_cpus_in_pool() primarily used the value from the
physical_procs_allocated_to_virtualization sysentry to represent the
cpus active in a pool. This is incorrect as that value represents
the physical processors in the system allocated to virtualization.

This patch changes the get_active_cpus_in_pool() function to only use
the pool_capacity sysentry for correctly determining the active cpus
in a pool.

Fixes: efa87d93231e ("lparstat: Correct presentation of some stats")

Signed-off-by: Luciano Chavez <chavez@us.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>