lsslot: don't report an error for an empty PHB list
If a system has no PHB's currently assigned to the system then the dlpar
node list for PHB types will be empty/NULL. This is not intrisically an
error and therefore we should return a successful return code on exit if
there is nothing to list.

Return zero in place of -1 if the PHB node list is empty.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>